requirements.txt
blinker==1.6.3
build==1.0.3
certifi==2024.2.2
click==8.1.7
colorama==0.4.6
elastic-transport==8.13.0
elasticsearch==8.13.1
Flask==3.0.0
importlib_metadata==7.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
packaging==23.2
pip-tools==7.3.0
pyproject_hooks==1.0.0
python-dotenv==1.0.0
tomli==2.0.1
urllib3==2.2.1
Werkzeug==3.0.1
zipp==3.18.2
